refactor: move fetch chunk enumeration to fetch module
